+++ Source/SwiftLintFramework/Extensions/Array+SwiftLint.swift
@@ -24,3 +24,27 @@ extension Array where Element: NSTextCheckingResult {
         return map { $0.range }
     }
 }
+
+extension Array where Element: Equatable {
+    var unique: [Element] {
+        var uniqueValues: [Element] = []
+        forEach { item in
+            if !uniqueValues.contains(item) {
+                uniqueValues += [item]
+            }
+        }
+        return uniqueValues
+    }
+}
+
+extension Array {
+    // swiftlint:disable:next line_length
+    func group<U: Hashable>(by transform: (Element) -> U) -> [U: [Element]] {
+        var dictionary: [U: [Element]] = [:]
+        for element in self {
+            let key = transform(element)
+            if case nil = dictionary[key]?.append(element) { dictionary[key] = [element] }
+        }
+        return dictionary
+    }
+}